version2

+++ passeplats.sh
@@ -2,6 +2,10 @@
 
 IN="/tmp/in"
 OUT="/tmp/out"
+LOCK="/tmp/out/lock"
+
+
+touch /tmp/out/lock
 
 if [ -d "!$IN" ]; then
 	exit 3
@@ -10,8 +14,12 @@ else
 	if [ -d "!$OUT" ]; then
 		mkdir /tmp/out 
 	fi
-
-	gzip -r /tmp/in/ 
-
-	cp /tmp/in/* /tmp/out/
+	if [ -f "$LOCK" ]; then
+		exit 22
+		echo "The script is already in use"
+	else
+		gzip -r /tmp/in/ 
+		cp /tmp/in/* /tmp/out/
+		rm /tmp/out/lock
+	fi
 fi
